+++ packages/featureserver/src/error.js
@@ -1,20 +1,21 @@
 const authenticationErrorResponse = require('../templates/errors/credentials-invalid.json')
 const authorizationErrorResponse = require('../templates/errors/unauthorized.json')
+const helpers = require('./helpers')
 
 /**
  * Respond with a authorization error response
  * @param {object} res express.js response object
  */
-function authorization (res) {
-  res.status(200).json(authorizationErrorResponse)
+function authorization (req, res) {
+  helpers.responseHandler(req, res, 200, authorizationErrorResponse)
 }
 
 /**
  * Respond with an authentication error response
  * @param {object} res express.js response object
  */
-function authentication (res) {
-  res.status(200).json(authenticationErrorResponse)
+function authentication (req, res) {
+  helpers.responseHandler(req, res, 200, authenticationErrorResponse)
 }
 
 module.exports = { authorization, authentication }