+++ packages/featureserver/test/error.js
@@ -3,9 +3,10 @@ const { error } = require('../src')
 
 describe('Error operations', () => {
   describe('authentication error', () => {
-    it('should return a status code 200 and error payload', () => {
+    it('without callback param, should return a status code 200 and error payload', () => {
       let statusCode
       let responsePayload
+      const req = { query: {} }
       const res = {
         status: function (code) {
           statusCode = code
@@ -15,7 +16,7 @@ describe('Error operations', () => {
           responsePayload = payload
         }
       }
-      error.authentication(res)
+      error.authentication(req, res)
       statusCode.should.equal(200)
       responsePayload.should.be.instanceOf(Object)
       responsePayload.should.have.property('error').instanceOf(Object)
@@ -24,12 +25,27 @@ describe('Error operations', () => {
       responsePayload.error.details.should.be.instanceOf(Array)
       responsePayload.error.details[0].should.equal('Invalid username or password.')
     })
+
+    it('should return a status code 200 and error payload', () => {
+      const req = { query: { callback: 'test' } }
+      const res = {
+        set: function (header, value) {
+          header.should.equal('Content-Type')
+          value.should.equal('application/javascript')
+        },
+        send: function (response) {
+          response.should.equal('test({"error":{"code":400,"message":"Unable to generate token.","details":["Invalid username or password."]}})')
+        }
+      }
+      error.authentication(req, res)
+    })
   })
 
   describe('authorization error', () => {
     it('should return a status code 200 and error payload', () => {
       let statusCode
       let responsePayload
+      const req = { query: {} }
       const res = {
         status: function (code) {
           statusCode = code
@@ -39,13 +55,27 @@ describe('Error operations', () => {
           responsePayload = payload
         }
       }
-      error.authorization(res)
+      error.authorization(req, res)
       statusCode.should.equal(200)
       responsePayload.should.be.instanceOf(Object)
       responsePayload.should.have.property('error').instanceOf(Object)
       responsePayload.error.should.have.property('code', 499)
       responsePayload.error.should.have.property('message', 'Token Required')
       responsePayload.error.details.should.be.instanceOf(Array)
+    })
+
+    it('with callback param, should return a status code 200 and error payload', () => {
+      const req = { query: { callback: 'test' } }
+      const res = {
+        set: function (header, value) {
+          header.should.equal('Content-Type')
+          value.should.equal('application/javascript')
+        },
+        send: function (response) {
+          response.should.equal('test({"error":{"code":499,"message":"Token Required","details":[]}})')
+        }
+      }
+      error.authorization(req, res)
     })
   })
 })