+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Authentication and authorization error messages now properly handled if request has `callback` query parameter
+
+### Added
+* Helper function to handle all success responses, with or without `callback` query parameter
+* `fullVersion: "10.5.1"` to info responses
+* 404 handling of `/MapServer`, `/MapServer/`, and ``/MapServer/:layer`
+
 ## [2.14.1] - 05-22-2018
 ### Fixed
 * Version bump due to NPM's inability to resolve last publish